+++ CampingSysteemAPI/Controllers/CampingRekeningController.cs
@@ -1,5 +1,6 @@
 ﻿using CampingSystem;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 
 namespace CampingSysteemAPI.Controllers
 {
@@ -56,6 +57,7 @@ namespace CampingSysteemAPI.Controllers
         {
             var existing = dal.GetCampingRekening(id);
             if (existing == null) return NotFound();
+            if (!existing.Reserveringen.IsNullOrEmpty()) return BadRequest("Deze camping rekening is verbonden aan reserveringen!");
 
             dal.DeleteCampingRekening(id);
             return NoContent();
